Index builder

+++ src/main/java/com/expleague/sensearch/donkey/crawler/CrawlerXML.java
@@ -2,6 +2,7 @@ package com.expleague.sensearch.donkey.crawler;
 
 import com.expleague.sensearch.ConfigJson;
 import com.expleague.sensearch.donkey.crawler.document.CrawlerDocument;
+import com.expleague.sensearch.donkey.crawler.document.WikiPage;
 import com.expleague.sensearch.donkey.crawler.document.XMLParser;
 import java.io.File;
 import java.io.FileInputStream;
@@ -106,39 +107,63 @@ public class CrawlerXML implements Crawler {
     }
   }
 
-
-
-    /*
-    public void makeZIP(int num) {
-        try {
-            FileInputStream fileInputStream = new FileInputStream(path.toString());
-            ZipInputStream zipInputStream = new ZipInputStream(fileInputStream);
-            ZipEntry zipEntry;
-            int cnt = 0;
-            XMLParser parser = new XMLParser();
-            while ((zipEntry = zipInputStream.getNextEntry()) != null) {
-                if (zipEntry.isDirectory()) {
-                    continue;
-                }
-                cnt++;
-                String fileName = "../WikiDocs/Mini_Wiki/";
-                Files.createDirectories(Paths.get(fileName));
-                String[] n = zipEntry.getName().split("/");
-                fileName += n[n.length - 1];
-
-                Files.copy(zipInputStream, Paths.get(fileName), StandardCopyOption.REPLACE_EXISTING);
-
-                zipInputStream.closeEntry();
-                File file = new File(fileName);
-                parser.parseXML(file);
-                if (cnt == num) {
-                    break;
-                }
-            }
-
-        } catch (IOException e) {
-            e.printStackTrace();
+  public void makeZIP(int num) {
+    try {
+      FileInputStream fileInputStream = new FileInputStream(path.toString());
+      ZipInputStream zipInputStream = new ZipInputStream(fileInputStream);
+      ZipEntry zipEntry;
+      int cnt = 0;
+      XMLParser parser = new XMLParser();
+      while ((zipEntry = zipInputStream.getNextEntry()) != null) {
+        if (zipEntry.isDirectory()) {
+          continue;
         }
-    }*/
+        cnt++;
+        String fileName = "../WikiDocs/Mini_Wiki/";
+        Files.createDirectories(Paths.get(fileName));
+        String[] n = zipEntry.getName().split("/");
+        fileName += n[n.length - 1];
+
+        if (!Files.exists(Paths.get(fileName))) {
+          zipInputStream.closeEntry();
+          continue;
+        }
+        Files.copy(zipInputStream, Paths.get(fileName), StandardCopyOption.REPLACE_EXISTING);
+
+        zipInputStream.closeEntry();
+        File file = new File(fileName);
+        WikiPage page = parser.parseXML(file);
+        boolean[] isOk = new boolean[1];
+        page.categories().forEach(category -> {
+          isOk[0] |= category.equals("Актёры России");
+          isOk[0] |= category.equals("Футболисты России");
+          isOk[0] |= category.equals("Писатели России по алфавиту");
+          isOk[0] |= category.equals("Поэты России");
+          isOk[0] |= category.equals("Актрисы России");
+          isOk[0] |= category.equals("Мастера спорта России международного класса");
+          isOk[0] |= category.contains("Правители");
+
+
+//          isOk[0] |= category.contains("Росс") && !category.contains("Википедия:") && !category.contains("ПРО:");
+          isOk[0] |= category.contains("Город") && !category.contains("Википедия:") && !category.contains("ПРО:");
+//          isOk[0] |= category.contains("Фильм") && !category.contains("Википедия:") && !category.contains("ПРО:");
+//          isOk[0] |= category.contains("фильм") && !category.contains("Википедия:") && !category.contains("ПРО:");
+        });
+
+//        if (!isOk[0]) {
+//          Files.delete(Paths.get(fileName));
+//        }
+        if (cnt % 100000 == 0) {
+          System.out.println(cnt);
+        }
+//        if (cnt == num) {
+//          break;
+//        }
+      }
+
+    } catch (IOException e) {
+      e.printStackTrace();
+    }
+  }
 }
